phonebook updates

+++ phonebook/src/main/java/com/coder2client/phonebook/ContactAddressRepository.java
@@ -2,5 +2,5 @@ package com.coder2client.phonebook;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactAddressRepository extends JpaRepository<ContactAddress, Integer> {
+public interface ContactAddressRepository extends JpaRepository<ContactAddress, Long> {
 }
